Move the blueprint saving logic to `re_viewer_blueprint`
